(BSR)[PRO] refactor: remove index file
