--- pro/src/ui-kit/form/shared/BaseFileInput/index.ts
@@ -1,1 +1,0 @@
-export { default as BaseFileInput } from './BaseFileInput'
