+++ pro/src/ui-kit/form/shared/index.ts
@@ -1,4 +1,3 @@
-export { BaseFileInput } from './BaseFileInput'
 export { default as BaseInput } from './BaseInput'
 export { default as BaseRadio } from './BaseRadio'
 export { default as FieldError } from './FieldError'
